Move all whiteSoldiers
I made the changes so now you can move all the white soldiers in the field. Besides that I them all in an ArrayList, so the methods won't be that stupid. Now I will try to place the array list in its class and do the same thing with the black army and also check to not put in a position that already is a soldier. :smile: